MOD overwrite results' creator with slotter for demo
usually the creator is the person who created the original event (the
metalslot). In case of a public slot this is probably ‘timeslot’. The
slotted is the user, who did a result. Because the iOS client is using
the ‘creator’ to determine to whom a slot belongs, we overwrite this
field until further changes.