IMP disable the ReSlot slot_type as a valid type in base_slot
I was able to make the specs runnable without that class. It’s not
necessary to have this slot_type for the production app, so it’s good
if it’s just not available. Unfortunately I had to change the custom
matcher, but not a big deal. I would like to do the same for std_slot.