MOD remove check which should always be true
because this class (self) has a method ‘reslots’, it will always
respond to it, so ```self.try(:reslots)``` should never return ```nil``
and so, should never be false. So I don’t think you need to use ‘try’
here. Or was it meant as a check if reslots exists? No reslots would
give [] and it would run the code but because each does not iterate
over an empty collection, it can be skipped on beforehand.